added no-network benchmark

+++ benchmarks/benchmark.py
@@ -36,20 +36,21 @@ def single_bench(clientName, clientNumber, functionID, iteration_count, benchsiz
     return data.DataSet(rtts, clientName, functionID, iteration_count, benchsize, clientNumber)
 
 def bench(clientName, serverName, clientNumbers, iteration_count, benchsize):
-    print("start server: " + serverName)
-    server = sp.Popen(["./benchmarks/" + serverName + "-server"], encoding="ascii", stdout=sp.PIPE, stderr=sp.PIPE, stdin=sp.PIPE)
+    if serverName != "":
+        print("start server: " + serverName)
+        server = sp.Popen(["./benchmarks/" + serverName + "-server"], encoding="ascii", stdout=sp.PIPE, stderr=sp.PIPE, stdin=sp.PIPE)
 
     results = data.Collection()
 
-    for functionID in range(1, 3):#5):
-
+    for functionID in range(1, 5):
         # warmup
         single_bench(clientName, clientNumbers[0], functionID, iteration_count, benchsize)
 
         for clientNumber in clientNumbers:
             results.add(single_bench(clientName, clientNumber, functionID, iteration_count, benchsize))
 
-    server.kill()
+    if serverName != "":
+        server.kill()
     return results
 
 def main():
@@ -61,14 +62,15 @@ def main():
 
     result_ctx = bench("ctx-net", "ctx-net", clientNumber, iteration_count, benchsize)
     result_mqtt = bench("mqtt", "mqtt", clientNumber, iteration_count, benchsize)
+    result_no_network = bench("no-network", "", clientNumber, iteration_count, benchsize)
 
-    merged_bench = result_mqtt.merge(result_ctx)
+    merged_bench = result_mqtt.merge(result_ctx).merge(result_no_network)
     #merged_bench.save("test.bench")
 
     #merged_bench = data.Collection()
     #merged_bench.load("test.bench")
 
-    #graph.plot_single_implementation(result_ctx, "ctx-net")
+    #graph.plot_single_implementation(result_no_network, "no-network")
     graph.plot_compare_implementations_scale_on_clients(merged_bench.filter_functionID(2))
 
 if __name__ == "__main__":
